Moved every class to its own file
