+++ index.html
@@ -4,7 +4,7 @@
 <title>DBIsnake</title>
 
 <script type="module">
-    import SnakeGame from './src/snake.js'
+    import SnakeGame from './src/SnakeGame.js'
 
     async function main() {
         let game = new SnakeGame(document.body, 3);
